Day 56_57_58_59_60 LLM & RAGs/LLM-RAG's Mini Projects/YouTube QA Bot/requirements.txt
streamlit
python-dotenv
langchain
langchain_community
youtube-transcript-api
langchain_huggingface
langchain_groq
langchain_openai
langchain_gemini
langchain_core

# Use Python 3.11 for better stability #